requirements.txt
Flask==2.3.2
boto3==1.26.84
requests==2.28.2
python-dotenv==0.21.0
pyjwt==2.6.0